Adds `HexAttribute` and associated property drawer.
`HexPropertyDrawer` will draw an integer type decorated with `[Hex]` represented as a hexadecimal string. The user can modify the hex value and it will re-serialize as expected. If the `0x` is removed from the field, or the user enters a decimal number, that number will be serialized as decimal, but still displayed as hexadecimal.

`HexAttribute` allows a manual setting of a minimum display width, adjusting how many potential 0s the hexadecimal string is padded with. By default, the the minimum width will be automatically determined based on the type of the field.