Add standard vertical spacing to some drawers.
A few drawers needed a little breathing room from the controls that followed them, in particular InfoBox and the help box drawn from `DrawDefaultPropertyAndHelpBox`. Added EditorGUIUtility.standardVerticalSpacing to the heights for those instances. As well, used EditorGUIUtility.standardVerticalSpacing for the height of controls inside a reorderable list, instead of the hardcoded value 4.