Update MinMaxSliderPropertyDrawer.cs
Because the min max slider is based on a vector2, the display for it works the same. Vector2s will show on the same line in the inspector as the tag only if the inspector is in widemode, otherwise, the width will double to allow the vector to be displayed on a line separate from the label. These changes just account for that and ensure it is displayed properly.

I typically keep my inspector as small as possible so the bug was really frustrating, I thought it was broken for the longest time until I saw this post : https://stackoverflow.com/questions/54814067/how-to-make-vector3-fields-behave-like-the-ones-from-transform-in-a-customeditor